vector-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
faiss-cpu==1.7.4
langchain==0.0.352
langchain-openai==0.0.2
openai==1.3.8
pydantic==2.5.2
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
python-multipart==0.0.6
PyPDF2==3.0.1
docx2txt==0.8